+++ clients/privacy-center/pages/api/fides-js.ts
@@ -113,6 +113,7 @@ export default async function handler(
         environment.settings.SERVER_SIDE_FIDES_API_URL ||
           environment.settings.FIDES_API_URL,
         environment.settings.DEBUG,
+        null,
         null
       );
     }